01_Workshop Day 1/app/tool_use_agent/requirements.txt
langgraph
langchain_anthropic
tavily-python
langchain_community
langchain_openai
qdrant-client 
pymupdf 
arxiv
tiktoken